requirements.txt
Django==4.2.4
djangorestframework==3.14.0
flake8==6.1.0
